soc: qcom: add modemsmem
Modems on some Google Pixel devices require extra information about the
SoC to be passed through a vendor-controlled Shared Memory (SMEM)
channel. This driver provides this information.

Without this driver, the modem might crash with the following message:

    [   17.262410] qcom-q6v5-mss 4080000.remoteproc: fatal error received: :0:[rf]EX:kernel:0:rf:171:PC=d949b6d4
    [   17.262454] remoteproc remoteproc0: crash detected in 4080000.remoteproc: type fatal error

Skimming through the device tree overlays, modemsmem seems to be used by
every Pixel chronologically between the Pixel 3 and Pixel 5a, inclusive.

The modem_flag field is not being initialized yet because the bootloader
doesn't always add the property in dtb (so it might not be necessary)
and because there are problems with how the proper value is being
retrieved in the original downstream driver.

Link: https://android.googlesource.com/kernel/msm/+/6bfa1545e98673df7399a58bfe9a0859b838b715/drivers/soc/qcom/modemsmem/modemsmem.c
Signed-off-by: Richard Acayan <mailingradian@gmail.com>
--
Discussion:

Where do these types of drivers go? This driver is probably providing
info to vendor code in the modem firmware, so it might not belong in
drivers/soc/qcom/.

Maybe it should go in another directory, like:

drivers/misc/
drivers/firmware/

I'm not too familiar with the kernel directory structure, so there might
be somewhere I missed.